Rename poolpair e2e to pool.pair e2e
